pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pangu-cli"
version = "0.1.0"
description = "盘古大模型平台管理 CLI"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "keyring>=24.0",
]

[project.scripts]
pangu = "pangu.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/pangu"]

[tool.hatch.build.targets.sdist]
include = ["src/pangu"]
